pyproject.toml
[project]
name = "hr-base-qa"
version = "0.1.0"
description = "An HR assistant service that combines RAG and agentic reasoning to extract external data and answer complex employee queries. Developed as part of a final project at the Higher School of Economics."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.1.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "latexbuild>=0.2.2",
    "openai>=1.74.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.2",
    "pydantic-settings>=2.8.1",
    "smolagents>=1.13.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.1.0",
    "pydantic-ai>=0.1.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.4",
]

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "EM",
  "FBT",
  "TRY003",
  "D1",
  "D203",
  "D213",
  "G004",
  "FA",
  "COM812",
  "ISC001",
  "RUF001",
]

[tool.ruff.lint.isort]
no-lines-before = ["standard-library", "local-folder"]
known-third-party = []
known-local-folder = []
lines-after-imports = 2

[tool.ruff.lint.extend-per-file-ignores]
"tests/*.py" = ["S101", "S311"]
